Added Platform.java.

+++ Q2/Crypto.java
@@ -1,11 +1,12 @@
 package Q2;
+import java.math.BigDecimal;
 
 public class Crypto {
     public int supply;
-    public double price;
+    public BigDecimal price;
     public String symbol;
 
-    public Crypto(String symbol, double price, int supply) { // Constructor. Will be used to create various coins if needed.
+    public Crypto(String symbol, BigDecimal price, int supply) { // Constructor. Will be used to create various coins if needed.
         this.symbol = symbol;
         this.price = price;
         this.supply = supply;
